refactor: example widget으로 분리
