+++ apps/client/src/views/example/index.tsx
@@ -1,14 +1,7 @@
-"use client";
-
-import { Header, Post } from "@/shared/ui";
-import { usePost } from "@repo/store/postProvider";
-import Mock from "@shared/mocks/data/evidenceMock.json";
-import { useRouter } from "next/navigation";
-import { post } from "@repo/types/evidences";
+import { Header } from "@/shared/ui";
+import ExampleWidget from "@/widgets/example/ui";
 
 export default function ExampleView() {
-  const { setPost } = usePost();
-  const R = useRouter();
   return (
     <div className="flex flex-col items-center">
       <Header />
@@ -16,21 +9,7 @@ export default function ExampleView() {
         <h1 className="text-tropicalblue-700 text-title4s sm:text-titleMedium text-left my-[2.38rem]">
           쓰기 예시 페이지
         </h1>
-        <div className="flex mt-[2.69rem] overflow-y-visible flex-wrap sm:justify-start justify-center w-full gap-[1.12rem]">
-          {Mock.map((data) => {
-            return (
-              <Post
-                onClick={() => {
-                  setPost(data as post);
-                  R.push(`/detail/${data.id}?example=${true}`);
-                }}
-                key={data.id}
-                data={data as post}
-                isExample={true}
-              />
-            );
-          })}
-        </div>
+        <ExampleWidget />
       </div>
     </div>
   );
